+++ breeze-core/src/main/kotlin/com/windea/breezeframework/core/annotations/ApiAnnotations.kt
@@ -4,32 +4,40 @@ import kotlin.annotation.AnnotationRetention.*
 
 /**需要显式使用的api。*/
 @MustBeDocumented
-@Retention(BINARY)
+@Retention(SOURCE)
 annotation class ExplicitUsageApi(
 	/**备注信息。*/
 	val message: String = ""
 )
 
-/**可能引起无限循环的api。*/
+/**需要选择性使用的api。*/
 @MustBeDocumented
-@Retention(BINARY)
-annotation class InfiniteLoopApi(
-	/**备注信息。*/
-	val message: String = ""
-)
-
-/**可能执行低性能操作的api。*/
-@MustBeDocumented
-@Retention(BINARY)
-annotation class LowPerformanceApi(
+@Retention(SOURCE)
+annotation class OptionalUsageApi(
 	/**备注信息。*/
 	val message: String = ""
 )
 
 /**难以或根本不可能实现的api。*/
 @MustBeDocumented
-@Retention(BINARY)
+@Retention(SOURCE)
 annotation class TrickImplementationApi(
 	/**备注信息。*/
 	val message: String = ""
 )
+
+///**可能引起无限循环的api。*/
+//@MustBeDocumented
+//@Retention(SOURCE)
+//annotation class InfiniteLoopApi(
+//	/**备注信息。*/
+//	val message: String = ""
+//)
+//
+///**可能执行低性能操作的api。*/
+//@MustBeDocumented
+//@Retention(SOURCE)
+//annotation class LowPerformanceApi(
+//	/**备注信息。*/
+//	val message: String = ""
+//)