+++ breeze-mapper/src/main/kotlin/com/windea/breezeframework/mapper/impl/JsonMapper.kt
@@ -3,6 +3,7 @@ package com.windea.breezeframework.mapper.impl
 import com.windea.breezeframework.core.domain.*
 import com.windea.breezeframework.core.extensions.*
 import com.windea.breezeframework.mapper.*
+import java.lang.reflect.*
 
 //DONE 生成json
 //TODO 解析json
@@ -117,6 +118,10 @@ class JsonMapper(
 
 
 	override fun <T> unmap(string: String, type: Class<T>): T {
+		TODO("not implemented")
+	}
+
+	override fun <T> unmap(string: String, type: Type): T {
 		TODO("not implemented")
 	}
 }