+++ breeze-serializer/src/main/kotlin/com/windea/breezeframework/serializer/extensions/SerializerExtensions.kt
@@ -10,23 +10,19 @@ import java.io.*
 //如果直接调用typeOf<T>()，会导致NotImplementedError，因此需要调用使用javaTypeOf<T>()
 
 /**从指定字符串读取指定类型的数据。*/
+@OptIn(ImplicitReflectionSerializer::class)
 inline fun <reified T : Any> Serializer.read(string: String): T {
 	return when(this) {
-		is KotlinSerializer<*> -> readForKotlinSerializer<T>(string)
+		is KotlinSerializer<*> -> delegate.parse(string)
 		else -> read(string, javaTypeOf<T>())
 	}
 }
 
 /**从指定文件读取指定类型的数据。*/
+@OptIn(ImplicitReflectionSerializer::class)
 inline fun <reified T : Any> Serializer.read(file: File): T {
 	return when(this) {
-		is KotlinSerializer<*> -> readForKotlinSerializer<T>(file.readText())
+		is KotlinSerializer<*> -> delegate.parse(file.readText())
 		else -> read(file, javaTypeOf<T>())
 	}
-}
-
-@PublishedApi
-@OptIn(ImplicitReflectionSerializer::class)
-internal inline fun <reified T : Any> KotlinSerializer<*>.readForKotlinSerializer(string: String): T {
-	return this.delegate.parse(string)
 }