+++ breeze-serializer/src/main/kotlin/com/windea/breezeframework/serializer/impl/BreezeSerializer.kt
@@ -1,0 +1,7 @@
+package com.windea.breezeframework.serializer.impl
+
+import com.windea.breezeframework.serializer.*
+
+/**由BreezeFramework实现的序列化器。*/
+@PublishedApi
+internal interface BreezeSerializer : Serializer