+++ breeze-serializer/src/main/kotlin/com/windea/breezeframework/serializer/impl/properties/JacksonPropertiesSerializer.kt
@@ -21,20 +21,20 @@ internal object JacksonPropertiesSerializer : PropertiesSerializer, JacksonSeria
 		return mapper.readValue(string, type)
 	}
 
-	override fun <T : Any> read(file: File, type: Class<T>): T {
-		return mapper.readValue(file, type)
-	}
-
-	override fun <T : Any> read(properties: Properties, type: Class<T>): T {
-		return mapper.readPropertiesAs(properties, type)
-	}
-
 	override fun <T : Any> read(string: String, type: Type): T {
 		return mapper.readValue(string, mapper.typeFactory.constructType(type))
 	}
 
+	override fun <T : Any> read(file: File, type: Class<T>): T {
+		return mapper.readValue(file, type)
+	}
+
 	override fun <T : Any> read(file: File, type: Type): T {
 		return mapper.readValue(file, mapper.typeFactory.constructType(type))
+	}
+
+	override fun <T : Any> read(properties: Properties, type: Class<T>): T {
+		return mapper.readPropertiesAs(properties, type)
 	}
 
 	override fun <T : Any> read(properties: Properties, type: Type): T {